Fix autosave not showing up in /journal/finish
